ESM (Mostly there!)
Not sure why I did this, especially all at once, but I converted as much of the library as I could over to ESM, as I would like to both see how and help out with how it works.

I'm working on an NBT editor project that I might like to utilize Ace with, and I thought having first-class TypeScript and ESM support would be something great to have here! Not sure if someone else has already done this, and this was all redundant, but I'm trying it out anyways. It's good for my own practice/skills/understanding. It's a bonus if the main project will be able to use these changes! That would be the ideal situation.

Ace has been a nice editor, I remember it back from when I was in High School AP CSP. Code.org used Ace, and I had always wondered how editing code in the browser worked, because you were writing code for the browser, in the browser. That was always very interesting to me. I wonder if that's what sparked the idea for stedit.app? :O

Being able to have tree-shaking, dynamic imports for language definitions, and built-in internal library checking would be a great thing for this project, it has been exceptionally awesome for my own projects, I can imagine how much more worthwhile it would be for bigger projects like these, where it really can shine.

I didn't work on any of the Test files yet, I just renamed them from `_test.js` to `.test.js` so I could distinguish them apart from the main source files.

VS Code refactoring was absolutely outstanding for this, it works great! You do have to go back and fix some of it's formatting changes though, as I didn't want to include other small changes in the commit that weren't specifically for ESM setup, and or ES6 class upgrades, those were the only two things for this work today.

This did take all way, probably about 4 separate sessions, had to be at least 8 hours? No way that's right, but it was probably more than that even. Why is this fun? I don't know, but I'm rolling with it!