Full Rename
Bringing my ESM-ed files over to TypeScript, I don't want to have the hassle of JSDoc types to start with. I do like them a lot, but for something so big as this project, having types right there is going to make it a little easier, at least hopefully.

Here was my little Bash script I wrote to do this for each folder.

```sh
for file in ./*.js
do
  name=$(basename -s .js "$file")

  # Debug
  echo $name
  mv $file "$name.ts"
done
```

I saved it in the root folder of the project, then just ran it after cd'ing into each directory that I wanted to rename.