Lib Types
I knew this project was going to be a lot to work on, but I think it's still more than I thought haha. Got a lot more done today, probably gonna relax a bit and take a break from it now. It was a nice thing to try doing, maybe I'll see if the current maintainers would like to extend any more of my progress at all.